requirements.txt
flask==1.1.2
gunicorn==20.0.4
flask_wtf==0.14.3
wtforms==2.3.3
redis==3.5.3
rq==1.7.0
flask_talisman==0.7.0
flask-session==0.3.2
Flask-JSGlue==0.3.1
flask-cors==3.0.9
Flask-Limiter==1.4
Flask-Caching==1.9.0
Flask-Security-Too==3.4.4
flask_mongoengine==1.0.0
flask_admin==1.5.6
bcrypt==3.2.0

matplotlib==3.5.1
networkx==2.5
xlrd==1.2.0
pubchempy==1.0.4
pyyaml==5.3.1
IPython==7.19.0
seaborn==0.11.0

pymed==0.8.9
crossrefapi==1.5.0
biopython==1.78
requests==2.27.1

keras==2.6.*
tables>=3.6.1,<4.0.0

pydot==1.4.1
palettable==3.3.0
unidecode==1.1.1
xmltodict==0.12.0
rq-scheduler==0.10.0

cirpy==1.0.2
natsort==7.1.1

bokeh==2.4.3
openpyxl==3.0.7
bioservices==1.8.3

jax-unirep==2.1.0
environs==9.5.0

tables==3.7.0

setuptools==59.5.0
pytorch-tree-lstm==0.1.3
torchmetrics==0.8.2
pytorch-lightning==1.6.3
typing-extensions==4.1.1